Add files via upload
The objective of this workshop was to empower and equip young African researchers with
computational modeling skills, enabling them to harness the full potential of cutting-edge
technologies in their respective fields. The workshop aimed to foster collaboration, innovation, and
the exchange of ideas among participants, thereby propelling research and development within
the region.